real space projectors is a mystery

+++ src/Band/add_nl_h_o_rs.cpp
@@ -18,7 +18,7 @@ void Band::add_nl_h_o_rs(K_point* kp__,
     auto rsp = parameters_.real_space_prj_;
     auto fft = parameters_.real_space_prj_->fft();
 
-    if (kappa__.size() < (2 * fft->size() + rsp->max_num_points_) * fft->num_fft_threads())
+    if (kappa__.size() < size_t(2 * fft->size() + rsp->max_num_points_) * fft->num_fft_threads())
     {
         TERMINATE("wrong size of work array");
     }
